feat(ci): pipeline now works, great
Made a few updates to the pipeline, now it passes:

- Deleted a solved comment
- Removed "export" command, it is not defined
- Bumped all github actions and node to latest

AFAIK, this is the first time the pipeline is running.
I have a pretty decent commit hook setup so that
has probably saved me from a lot of trouble, but
this is a good step.

Another goal was to auto-deploy the build to S3,
but when I looked at credentials etc that seems
like a bit of an effort.

Also, I'd want to ideally do that using terraform
to stick with this IaC (infrastructure as code)
idea, but that'll take some more effort.

So for now: Pushing this into `main`, with a green
tick on the pipeline.